backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6 
langchain==0.3.1 
langchain-google-genai==2.1.5 
langchain-community==0.3.1 
sentence-transformers==3.1.1 
faiss-cpu==1.8.0 
python-multipart==0.0.9 
python-dotenv==1.0.1
fastapi==0.115.4 
uvicorn==0.32.0 
python-multipart==0.0.12 
langchain-google-genai==2.0.1 
python-dotenv==1.0.1 
langchain==0.3.4 
langchain-community==0.3.3 
faiss-cpu==1.9.0 
sentence-transformers==3.2.1 
PyPDF2==3.0.1